Move execution context into -lib-data
