+++ tracdap-libs/tracdap-lib-data/src/main/java/org/finos/tracdap/common/data/DataContext.java
@@ -16,7 +16,6 @@
 
 package org.finos.tracdap.common.data;
 
-import org.finos.tracdap.common.concurrent.ExecutionContext;
 import io.netty.util.concurrent.OrderedEventExecutor;
 import org.apache.arrow.memory.BufferAllocator;
