+++ tracdap-libs/tracdap-lib-data/src/main/java/org/finos/tracdap/common/storage/CommonFileStorage.java
@@ -17,7 +17,7 @@
 package org.finos.tracdap.common.storage;
 
 import org.finos.tracdap.common.concurrent.Flows;
-import org.finos.tracdap.common.concurrent.IExecutionContext;
+import org.finos.tracdap.common.data.IExecutionContext;
 import org.finos.tracdap.common.config.ConfigHelpers;
 import org.finos.tracdap.common.data.IDataContext;
 import org.finos.tracdap.common.exception.ETrac;
