+++ tracdap-libs/tracdap-lib-data/src/test/java/org/finos/tracdap/common/storage/local/LocalStorageReadWriteTest.java
@@ -19,7 +19,6 @@ package org.finos.tracdap.common.storage.local;
 import io.netty.util.concurrent.DefaultEventExecutor;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import org.apache.arrow.memory.RootAllocator;
-import org.finos.tracdap.common.concurrent.ExecutionContext;
 import org.finos.tracdap.common.data.DataContext;
 import org.finos.tracdap.common.storage.StorageReadWriteTestSuite;
 import org.finos.tracdap.common.storage.IStorageManager;
