+++ tracdap-libs/tracdap-lib-data/src/test/java/org/finos/tracdap/common/storage/local/LocalStorageStabilityTest.java
@@ -16,7 +16,7 @@
 
 package org.finos.tracdap.common.storage.local;
 
-import org.finos.tracdap.common.concurrent.ExecutionContext;
+import org.finos.tracdap.common.data.ExecutionContext;
 import org.finos.tracdap.common.data.DataContext;
 import org.finos.tracdap.common.storage.StorageStabilityTestSuite;
 import org.finos.tracdap.common.storage.IStorageManager;
