+++ tracdap-plugins/aws-storage/src/test/java/org/finos/tracdap/plugins/aws/storage/S3StorageOperationsTest.java
@@ -21,7 +21,7 @@ import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import org.apache.arrow.memory.BufferAllocator;
 import org.apache.arrow.memory.RootAllocator;
-import org.finos.tracdap.common.concurrent.ExecutionContext;
+import org.finos.tracdap.common.data.ExecutionContext;
 import org.finos.tracdap.common.data.DataContext;
 import org.finos.tracdap.common.storage.StorageOperationsTestSuite;
 import org.junit.jupiter.api.*;
