+++ tracdap-plugins/aws-storage/src/test/java/org/finos/tracdap/plugins/aws/storage/S3StorageReadWriteTest.java
@@ -19,7 +19,7 @@ package org.finos.tracdap.plugins.aws.storage;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.nio.NioEventLoopGroup;
 import org.apache.arrow.memory.BufferAllocator;
-import org.finos.tracdap.common.concurrent.ExecutionContext;
+import org.finos.tracdap.common.data.ExecutionContext;
 import org.finos.tracdap.common.data.DataContext;
 import org.finos.tracdap.common.storage.StorageReadWriteTestSuite;
