+++ tracdap-services/tracdap-svc-data/src/test/java/org/finos/tracdap/svc/data/api/DataRoundTripTest.java
@@ -18,9 +18,9 @@ package org.finos.tracdap.svc.data.api;
 
 import org.apache.arrow.vector.ipc.ArrowFileWriter;
 import org.finos.tracdap.api.*;
-import org.finos.tracdap.common.concurrent.ExecutionContext;
+import org.finos.tracdap.common.data.ExecutionContext;
 import org.finos.tracdap.common.concurrent.Flows;
-import org.finos.tracdap.common.concurrent.IExecutionContext;
+import org.finos.tracdap.common.data.IExecutionContext;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.plugin.PluginManager;
 import org.finos.tracdap.common.util.ResourceHelpers;
