+++ tracdap-services/tracdap-svc-data/src/test/java/org/finos/tracdap/svc/data/api/FileRoundTripTest.java
@@ -17,10 +17,10 @@
 package org.finos.tracdap.svc.data.api;
 
 import org.finos.tracdap.api.*;
-import org.finos.tracdap.common.concurrent.ExecutionContext;
+import org.finos.tracdap.common.data.ExecutionContext;
 import org.finos.tracdap.common.concurrent.Flows;
 import org.finos.tracdap.common.concurrent.Futures;
-import org.finos.tracdap.common.concurrent.IExecutionContext;
+import org.finos.tracdap.common.data.IExecutionContext;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.plugin.PluginManager;
 import org.finos.tracdap.metadata.CopyStatus;
