+++ tracdap-services/tracdap-svc-data/src/test/java/org/finos/tracdap/svc/data/api/TenantSeparationTest.java
@@ -18,8 +18,8 @@ package org.finos.tracdap.svc.data.api;
 
 import org.finos.tracdap.api.TracDataApiGrpc;
 import org.finos.tracdap.api.TracMetadataApiGrpc;
-import org.finos.tracdap.common.concurrent.ExecutionContext;
-import org.finos.tracdap.common.concurrent.IExecutionContext;
+import org.finos.tracdap.common.data.ExecutionContext;
+import org.finos.tracdap.common.data.IExecutionContext;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.metadata.MetadataUtil;
 import org.finos.tracdap.common.plugin.PluginManager;
